modules/kafka-consumer/requirements.txt
kafka-python
marshmallow
psycopg2-binary